refactor(main|SettingTab): 重构资产状态管理，将相关逻辑移动到独立模块，优化代码结构 #57
